ai_agency_platform/services/auth-service/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
emails
psycopg2-binary
email-validator
bcrypt==3.2.0
